requirements.txt
flask
requests
beautifulsoup4
python-dotenv
pydantic
pytest
black